refactor: move interfaces from generics to TestxApi
