--- src/generics.ts
@@ -1,9 +1,0 @@
-export interface ImportData {
-  [table: string]: {
-    [column: string]: unknown;
-  }[];
-}
-
-export interface DatabaseSchema {
-  [table: string]: string[];
-}
